Add caching to connection
